requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.30.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
MouseInfo==0.1.3
mss==10.1.0
narwhals==2.3.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
